refactor: move doctor setup functions to testHelpers.js
