+++ tests/routes/patientTemplate/addFieldToPatientTemplate.test.js
@@ -1,17 +1,21 @@
-const { createDoctor, deleteDoctor } = require('../doctorSetup')
 const axios = require('axios')
-const { BASE_URL } = require('../../jest.setup')
+const { BASE_URL, getAuthToken } = require('../../jest.setup')
+const { createTestDoctor, deleteUser } = require('../../testHelpers')
 
 let doctorId, headers
 
 beforeAll(async () => {
-  const setup = await createDoctor()
-  doctorId = setup.doctorId
-  headers = setup.headers
+  const doctor = await createTestDoctor()
+  doctorId = doctor.id
+  headers = {
+    'Content-Type': 'application/json',
+    Authorization: `Bearer ${getAuthToken()}`,
+    Origin: 'http://localhost'
+  }
 })
 
 afterAll(async () => {
-  await deleteDoctor(doctorId, headers)
+  await deleteUser(doctorId)
 })
 
 describe('Create Patient Template Tests', () => {
